# api/openapi/qavia.yaml
openapi: 3.1.0
...
info:
  title: Qavia API
  version: 0.1.0
  description: |
    Internal Hyscaler platform that acts as an AI QA engineer.

    Every mutation goes through this API. The web app holds no database
    credentials and no queue access, so this is the single place authorisation is
    enforced.
  contact:
    name: Hyscaler
    email: max@hyscaler.com
servers:
  - url: /
    description: Same origin as the API process
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/projects/{projectID}/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: A project's jobs
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of jobs, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPage"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      operationId: submitJob
      tags: [jobs]
      summary: Submit a job chain
      description: |
        Returns 202 with a job reference and never blocks: the user may close the
        browser and be notified when the chain finishes (FR-1.6, NFR-1).

        Everything checkable is checked here rather than inside a worker twenty
        minutes later: the project is not archived, the chain exists, an AI chain
        has a configured provider, and the project is approved for external AI if
        the assigned provider is not local.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitJobRequest"
            example:
              chain: noop
      responses:
        "202":
          description: Accepted. Subscribe to the job rather than waiting.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobReference"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/jobs/{jobID}/cancel:
    post:
      operationId: cancelJob
      tags: [jobs]
      summary: Cancel a job
      description: |
        Cancellation is cooperative: the handler's context is cancelled and it
        stops between units of work, so nothing is killed mid-write. A job that has
        already finished returns `job_not_cancelable`.
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "200":
          description: Cancellation requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Already finished.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: job_not_cancelable
                message: This job is succeeded and can no longer be cancelled.
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/jobs/{jobID}/events:
    get:
      operationId: streamJobEvents
      tags: [jobs]
      summary: Live job status and event log
      description: |
        Server-sent events. The current state is sent on connect and then deltas,
        so a late joiner is never left with a blank screen.

        A comment heartbeat every 15 seconds keeps proxies from closing an idle
        stream. `Last-Event-ID` resumes after a reconnect from the last event the
        client saw, so a dropped connection does not leave a gap in the log.

        Fan-out is one database listener per process serving many browsers, never
        one connection per viewer.
      parameters:
        - $ref: "#/components/parameters/JobID"
        - name: Last-Event-ID
          in: header
          required: false
          description: The id of the last event received. Sent automatically by an EventSource on reconnect.
          schema:
            type: string
            maxLength: 32
      responses:
        "200":
          description: An event stream of `status` and `event` messages.
          content:
            text/event-stream:
              schema:
                type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qavia_session
      description: |
        Server-side session. Sessions live in Postgres rather than in a JWT, so
        revocation is immediate: an offboarded employee's access ends when the row
        is deleted, with no token TTL to wait out.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Every list has a hard maximum; there is no unbounded list.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous response's nextCursor. Do not construct one.
      schema:
        type: string
        maxLength: 512
...
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    JobID:
      name: jobID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Supported on every mutating endpoint a CI system or webhook can retry.
        Replaying the same key with the same body returns the original result;
        reusing it with a different body is a conflict.
      schema:
        type: string
        maxLength: 255
  schemas:
    Error:
      type: object
      description: |
        The single error shape. Produced only by the mapper in
        internal/platform/httpx, so every failure looks the same.
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable and machine readable. Branch on this, never on message. Every
            value is catalogued in docs/error-codes.md.
          example: target_host_not_allowed
        message:
          type: string
          description: Safe to show a user, and says what to do next.
          example: Host "api.example.test" is not on this project's allowlist. Add it in Settings, Project, Targets.
        details:
          type: object
          additionalProperties: true
          description: Field-level context for inline rendering. Never carries a secret.
          example:
            host: api.example.test
        incidentId:
          type: string
          description: |
            Present only on a 500. The handle between what the user saw and the
            full wrapped cause in the log, which is never returned.
          example: 9f2c1a7e
...
    NextCursor:
      type: [string, "null"]
      description: Pass as cursor to fetch the next page. Null or absent means the last page.
      example: eyJjIjoiMjAyNi0wOC0xMFQwOToxNDoyMloifQ
    JobReference:
      type: object
      description: |
        Returned with 202 by every long operation. The client subscribes to the
        job rather than waiting: the user may close the browser and be notified
        when the chain finishes.
      required: [jobId, status]
      properties:
        jobId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/JobStatus"
        eventsUrl:
          type: string
          description: SSE endpoint for live status and the event log.
          example: /api/v1/jobs/6f1c.../events
      example:
        jobId: 6f1c9d2e-6b1a-4f7a-9c3e-2b8d5a4c1e77
        status: queued
        eventsUrl: /api/v1/jobs/6f1c9d2e-6b1a-4f7a-9c3e-2b8d5a4c1e77/events
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
...
    JobChain:
      type: string
      description: |
        Chains are declared in one place on the server. A handler never calls the
        next handler directly, so every stage is independently retryable and
        independently observable.
      enum: [noop, ai_smoke, ingest, generate, execute, analyse]
    SubmitJobRequest:
      type: object
      required: [chain]
      properties:
        chain:
          $ref: "#/components/schemas/JobChain"
        artifactId:
          type: string
          format: uuid
          description: The input to work from, where the chain needs one.
    JobEventLevel:
      type: string
      enum: [debug, info, warn, error]
    JobEvent:
      type: object
      description: One line of the live event log. `id` doubles as the SSE Last-Event-ID.
      required: [id, level, message, at]
      properties:
        id:
          type: string
          description: Ascending. Pass the last one you saw as Last-Event-ID to resume.
          example: "4821"
        level:
          $ref: "#/components/schemas/JobEventLevel"
        message:
          type: string
        at:
          type: string
          format: date-time
    Job:
      type: object
      required: [id, type, status, progress, attempts, maxAttempts, queuedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: [string, "null"]
          format: uuid
        type:
          type: string
          description: The handler that runs this stage.
          example: noop.stage1
        chain:
          $ref: "#/components/schemas/JobChain"
        status:
          $ref: "#/components/schemas/JobStatus"
        progress:
          type: integer
          minimum: 0
          maximum: 100
        attempts:
          type: integer
        maxAttempts:
          type: integer
        error:
          type: [string, "null"]
          description: Set once retries are exhausted. Never carries a credential.
        parentJobId:
          type: [string, "null"]
          format: uuid
        stages:
          type: array
          description: Present on a parent job, so one request describes the whole chain.
          items:
            $ref: "#/components/schemas/Job"
        events:
          type: array
          description: The most recent event lines. The SSE stream carries the rest.
          items:
            $ref: "#/components/schemas/JobEvent"
        queuedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
    JobPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        nextCursor:
          $ref: "#/components/schemas/NextCursor"
...
  responses:
    BadRequest:
      description: The request is structurally valid but not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: validation_failed
            message: Email is not valid.
            details:
              field: email
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthenticated
            message: Sign in to continue.
    Forbidden:
      description: Authenticated, but not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: role_required
            message: This action requires the Admin role.
    NotFound:
      description: |
        Not found, or not visible to this caller. The two are deliberately
        indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: project_not_found
            message: Project not found.
    Conflict:
      description: The request conflicts with current state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: project_archived
            message: This project is archived and is read-only. Unarchive it to make changes.
...
    InternalError:
      description: |
        An unexpected failure. Carries an incident ID and nothing else; the cause
        is in the structured log against that ID.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: internal_error
            message: Something went wrong on our side.
            incidentId: 9f2c1a7e